+++ data/types/upload/upload.go
@@ -94,11 +94,13 @@ func NewUpload(parser data.ObjectParser) *Upload {
 }
 
 func ParseUpload(parser data.ObjectParser) *Upload {
-	datum := NewUpload(parser)
-	if datum == nil {
+	if parser.Object() == nil {
 		return nil
 	}
 
+	_ = parser.ParseString("type")
+
+	datum := New()
 	datum.Parse(parser)
 	return datum
 }
@@ -155,16 +157,16 @@ func (u *Upload) Validate(validator structure.Validator) {
 		validator.String("dataState", u.DataState).OneOf(States()...)
 	}
 
-	validator.StringArray("deviceManufacturers", u.DeviceManufacturers).Exists().NotEmpty().EachNotEmpty().EachUnique()
-	validator.String("deviceModel", u.DeviceModel).Exists().NotEmpty()               // TODO: Some clients USED to send ""; requires DB migration
-	validator.String("deviceSerialNumber", u.DeviceSerialNumber).Exists().NotEmpty() // TODO: Some clients STILL send "" via Jellyfish; requires fix & DB migration
-	validator.StringArray("deviceTags", u.DeviceTags).Exists().NotEmpty().EachOneOf(DeviceTags()...).EachUnique()
+	validator.StringArray("deviceManufacturers", u.DeviceManufacturers).NotEmpty().EachNotEmpty().EachUnique()
+	validator.String("deviceModel", u.DeviceModel).NotEmpty()               // TODO: Some clients USED to send ""; requires DB migration
+	validator.String("deviceSerialNumber", u.DeviceSerialNumber).NotEmpty() // TODO: Some clients STILL send "" via Jellyfish; requires fix & DB migration
+	validator.StringArray("deviceTags", u.DeviceTags).NotEmpty().EachOneOf(DeviceTags()...).EachUnique()
 
 	if validator.Origin() <= structure.OriginInternal {
 		validator.String("state", u.State).OneOf(States()...)
 	}
 
-	validator.String("timeProcessing", u.TimeProcessing).Exists().OneOf(TimeProcessings()...) // TODO: Some clients USED to send ""; requires DB migration
+	validator.String("timeProcessing", u.TimeProcessing).OneOf(TimeProcessings()...) // TODO: Some clients USED to send ""; requires DB migration
 	validator.String("version", u.Version).LengthGreaterThanOrEqualTo(VersionLengthMinimum)
 }
 
@@ -177,7 +179,7 @@ func (u *Upload) Normalize(normalizer data.Normalizer) {
 
 	if normalizer.Origin() == structure.OriginExternal {
 		if u.UploadID == nil {
-			u.UploadID = pointer.FromString(data.NewSetID())
+			u.UploadID = u.ID
 		}
 	}